clients/DiscordBot/pyproject.toml
[tool.poetry]
name = "discord-club-bot"
version = "0.1.0"
description = "Discord bot LVP with QR code generation via slash commands"
authors = ["Austin Wagner <austinwagner@msn.com>"]
readme = "README.md"
packages = [
  { include = "clubbot", from = "src" },
  { include = "scripts" },
]
include = ["src/clubbot/py.typed"]

[tool.poetry.dependencies]
python = "^3.11"
"discord.py" = "^2.4.0"
Pillow = "^10.1.0"
python-dotenv = "^1.0.0"
rich = "^13.7.1"
httpx = "^0.28.1"
redis = "^5.2.0"
rq = "^2.0.0"
monorepo-guards = { path = "../../libs/monorepo_guards", develop = true }
platform-core = { path = "../../libs/platform_core", develop = true }
platform-discord = { path = "../../libs/platform_discord", develop = true }
platform-workers = { path = "../../libs/platform_workers", develop = true }

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.0"
pytest-asyncio = "^1.3.0"
pytest-cov = "^7.0.0"
pytest-xdist = "^3.6.1"
ruff = "^0.14.4"
black = "^25.11.0"
mypy = "^1.13.0"
types-Pillow = "^10.2.0.20240822"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

# Ruff configuration (linter + formatter)
[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "scripts", "tests"]
exclude = [".venv"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
    "B",  # bugbear
    "BLE",# blind except
    "UP", # pyupgrade
    "N",  # pep8-naming
    "C4", # flake8-comprehensions
    "SIM",# flake8-simplify
    "RET",# flake8-return
    "C90",# mccabe complexity
    "RUF",# Ruff-specific rules
]
ignore = []
fixable = ["ALL"]

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"typing.Any" = { msg = "Do not use typing.Any; prefer precise types or Protocols/TypedDicts." }
"typing.cast" = { msg = "Do not use typing.cast; prefer adapters or precise types." }

[tool.ruff.lint.per-file-ignores]
# No per-file ignores: repository is kept pristine. All violations must be fixed.

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
disallow_subclassing_any = true
disallow_any_generics = true
disallow_any_unimported = true
disallow_any_expr = true
disallow_any_decorated = true
disallow_any_explicit = true
no_implicit_optional = true
check_untyped_defs = true
no_implicit_reexport = true
show_error_codes = true
pretty = true
files = ["src", "tests", "scripts"]
mypy_path = ["src"]
explicit_package_bases = true

[tool.coverage.run]
source = ["src", "scripts"]
omit = []
branch = true

[tool.coverage.report]
precision = 2
show_missing = true
fail_under = 100

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v -n auto --dist loadscope"
filterwarnings = [
  "ignore:'audioop' is deprecated:DeprecationWarning:discord.player",
]
